Use `Messager` from `processingEnv` for logging
As I started running experiments where more text was printed to stdout,
it seemed as if Gradle or the Gradle Daemon began suppressing this
output. I am not exactly sure why. One work-around was to simply restart
the Gradle daemon, though that was a bit annoying.

This commit changes the `TypesProc` to instead use standard annotation
processing logging methodology: `javax.annotation.processing.Messager`.
This instance is made available to the annotation processor via the
`ProcessingEnvironment` instance.