Add a method with body to `:types:driver`
This helps demonstrate that when a `TypeElement` is written out, the
method bodies are not printed along with the interface information.
In general, the Elements API is not designed to reveal information
from method bodies. (IIRC, more information on this is available
somewhere in the Annotation Processing API's Javadoc.)

However, as we can see over in the `:expr` experiment, method body
information *is* available via the Trees API.