Is this field set once processing is over? No.
This commit extends the experiment in `:types` which was added in the
last commit (0a344360). It checks to see whether the
`JCMethodInvocation.type` field is set once the annotation processing
is flagged to be over (i.e. once `roundEnv.processingOver()` is true).
The answer, once again, appears to be *no*. This field must be set
at some still-later point in compilation.

Evidence for this was provided by performing two modifications:

1. Initialize a map of `CompilationUnitTree` instances *before*
processing is over.
2. Check both before and after processing is over whether the relevant
fields are set, both times checking a `CompilationUnitTree` stored in
this map.