Is this field set once analysis phase is over? Yes!
The last two commits developed an experiment in `:types` to see whether
the `JCMethodInvocation.type` field is set at some point during
annotation processing. It was found that it was not set.

In this commit, the experiment was extended to check this field again
after *analysis* phase of compilation was complete, and it was found
that the field is set at this time.

The ability to perform this check at this phase of compilation is not
provided via the standard annotation processing APIs. Instead, this
experiment used a `javac`-specific API which I just found in
`com.sun.source` based on `JavacTask`.

Note that the experiment was still implemented using the annotation
processing API: the processor was used to register a callback which was
invoked by the compiler after analysis.