Use `java-compile-utils` to invoke sub-compilation
This successfully performs a sub-compilation using file manager
locations inherited from the initial compilation. Thus some of the main
compilation configuration performed by Gradle can be inherited by the
sub-compilation.

The `:sub:proc` subproject has been extended beyond just a stub. It
now implements an annotation processor which uses a
`me.dwtj.java.compiler.runner.CompilationTaskBuilder` to perform a
sub-compilation within this annotation processor.

As can be seen in `noteClassPathOfFileManagerFromToolProvider()`, the
standard file manager provided via the `ToolProvider` is not initialized
with very much useful location information, i.e. locations like
`CLASS_PATH` and `CLASS_OUTPUT` are essentially un-initialized.

I wanted for the sub-compilation to somehow inherit most of these
locations from the compilation in which the annotation processor is
running. It appears that the standard APIs do not allow this, so I used
reflection to extract the appropriately-initialized file manager
instance from the `Filer` provided to the annotation processor; I then
copied most of these locations over to the file manager to be used in
the sub-compilation.

As can be seen in `trySubCompilation()`, the annotation processor now
triggers a sub-compilation which places the resulting class file in
a temporary directory.

An important detail that is required to make this work is the explict
setting of `-sourcepath` in `sub/driver/build.gradle`. By setting this
`javac` command line argument directly, the sourcepath of the file
manager from the `Filer` will have this absolute directory; the
sub-compilation's file manager will therefore inherit this directory.
Without setting this command line argument, Gradle sets the sourcepath
to an empty directory, from which the desired class cannot be loaded.
Check the following links for more info:

    - http://blog.ltgt.net/most-build-tools-misuse-javac/
    - https://discuss.gradle.org/t/gradle-should-pass-sourcepath-to-javac-by-default-to-avoid-false-positives/2288