Copy sub-compiled class to proc's `Filer`
This commit adds a simple experiment to `:sub` to see whether by adding
a sub-compiled class to an annotation processor's `Filer` it might
prevent the compilation of that class by the initial compilation. It
appears that the answer is *no* if filing is performed after processing
is over but *yes* if filing is performed before this.

This was done by using the debugger to pause the compilation between
when the `Filer` has copied the class file and before the annotation
processing phase has ended, and then checking file creation/modification
timestamps of

   sub/driver/build/classes/main/me/dwtj/ex/openjdk/langtools/sub/SubDriverClass.class

That manually filing the class during annotation processing prevents
compilation could be further validated by using a javac compilation task
listener to see whether GENERATE begin/end events are emitted for this
class.

It may be the case that the above-described behavior may not occur in
other environments (e.g. Eclipse).